// src/modules/causal-knowledge/graphicalCausalKnowledgeBase.d.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND.
 * Instead, modify the source JSONSchema file,
 * and run `npm run generate` to regenerate this file.
 */

export type Id = number;

/**
 * Graphcial represenation of causal knowledge base.
 */
export interface GraphicalCausalKnowledgeBase {
  $schema?: string;
  apiVersion: "graphical/v1";
  atoms: Atom[];
  operators: Conjunction[];
  connections: Connection[];
}
export interface Atom {
  id: Id;
  name: string;
  description: string;
  assumption?: 1 | 2 | 3 | 4 | 5;
  position: Position;
}
export interface Position {
  x: number;
  y: number;
}
export interface Conjunction {
  id: Id;
  type: "conjunction";
  position: Position;
}
export interface Connection {
  id: ConnectionId;
  negated: boolean;
}
export interface ConnectionId {
  sourceId: Id;
  targetId: Id;
}